Tests use test database. Move tests for routes into routes directory
