+++ server/package.json
@@ -10,7 +10,7 @@
     "server": "nodemon server",
     "client": "npm start --prefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
-    "test": "jest"
+    "test": "DATABASE_NAME=shopfronttest DATABASE_USER=ellej1 DATABASE_PASSWORD=hello DATABASE_HOST=localhost DATABASE_PORT=5432 jest"
   },
   "repository": {
     "type": "git",
