+++ server/routes/__tests__/products_test.js
@@ -1,6 +1,5 @@
 const request = require('supertest');
-const app = require('../app');
-require('custom-env').env(true);
+const app = require('../../app');
 
 describe('GET /api/products', () => {
 	test('responds with json containing product details, and 200 status', function (done) {
