+++ server/routes/__tests__/users_test.js
@@ -1,19 +1,16 @@
 const request = require('supertest');
-const usersTest = require('../database/users');
-const app = require('../app');
-require('custom-env').env(true);
+const usersTest = require('../../database/users');
+const app = require('../../app');
 
 beforeEach(() => {
   usersTest.clearTestUserTable()
 });
 
 describe('GET /users/someroute', () => {
-  console.log(process.env.DATABASE_NAME)
   test('responds with 404 status for an incorrect route', function (done) {
     request(app)
       .get('/users/someroute')
       .expect(404, done)
-
   });
 })
